Add support for return dataset files
Introduces an argument for optionally returning the list of files used
to construct the dataset. While useful in some circumstances, note that
this does not work properly if the dataset is later shuffled, batched,
and so forth. Mostly useful in scenarios where we want to e.g. analyze
trained models through Monte Carlo Dropout, LRP, etc.