Monte Carlo Dropout (#6)
* Add support for returning dataset files

Introduces an argument for optionally returning the list of files used
to construct the dataset. While useful in some circumstances, note that
this does not work properly if the dataset is later shuffled, batched,
and so forth. Mostly useful in scenarios where we want to e.g. analyze
trained models through Monte Carlo Dropout, LRP, etc.

* Implement Monte Carlo Dropout support

Adds support for Monte Carlo Dropout by implementing a custom Dropout
layer that can be forced to always be enabled, even during inference.
This was necessary since the built in Dropout layer is only active
during model training, and is not readily "forced".

* Implement Monte Carlo Dropout simulations

Runs `N` Monte Carlo Dropout simulations on a trained model, and saves
the resulting predictions and accuracies for all simulations. The saved
results can be used for further analysis in e.g. Jupyter Notebook.

* Un-force Dropout on trained models

Not necessary to always force dropout, since this can be enabled as
required during inference when performing model analysis. Because the
`force` argument was set to true, we were forced to actively disable
dropout during inference in all "normal" use cases.

* Implement basic analysis report

* Perform other minor tweaks and changes to support Monte Carlo Dropout